requirements.txt
flask==2.3.2
flask-socketio==5.8.0
flask-cors==4.0.0
openai==0.27.8
python-dotenv==1.0.0
pydub==0.25.1
gunicorn==20.1.0
websockets==11.0.3
reportlab==4.0.4
python-docx==0.8.11
eventlet==0.33.3
google-cloud-secret-manager==2.16.4
